testing/python/requirements.txt
absl-py
grpclib==0.3.1
